Perfect graphs and add parameter correlation
This Matlab file is now a function that is able to plot the effect of the volume fraction and orientation of the secondary fiber with respect to a certain pulling direction (here: the direction of the primary fiber). The function is also able to mark the intersection line between the graphs and give the equation of this line in the form of a 1st-order polynomial in a string.